src/parser/docx_parser.py
from typing import Dict, List, Any, Optional, Tuple
from docx import Document as DocxDocument
from docx.document import Document as DocxDocumentClass
from docx.table import Table as DocxTable
from docx.text.paragraph import Paragraph as DocxParagraph

from ..models import Document, Paragraph, Image, Table

class DocxParser:
    """
    Parser for DOCX files.
    Converts DOCX elements into the document model structure.
    """
    
    def __init__(self):
        self.image_counter: int = 1  # Counter for generating image filenames
        self.document: Optional[Document] = None
        self.docx: Optional[DocxDocumentClass] = None
        
    def parse_document(self, docx_path: str) -> Document:
        """
        Parse a DOCX file and return a Document object.
        
        Args:
            docx_path (str): Path to the DOCX file
            
        Returns:
            Document: Parsed document object
        """
        pass
        
    def _parse_paragraph(self, paragraph: DocxParagraph) -> Paragraph:
        """
        Parse a paragraph element from docx
        
        Args:
            paragraph: A docx paragraph object
            
        Returns:
            Paragraph: The parsed Paragraph object
        """
        pass
        
    def _extract_image(self, paragraph: DocxParagraph) -> Optional[Image]:
        """
        Extract image from a paragraph
        
        Args:
            paragraph: A docx paragraph that might contain an image
            
        Returns:
            Optional[Image]: The extracted image or None
        """
        pass
        
    def _parse_table(self, table: DocxTable) -> Table:
        """
        Parse a table element from docx
        
        Args:
            table: A docx table object
            
        Returns:
            Table: The parsed Table object
        """
        pass
    
    def _detect_heading_level(self, paragraph: DocxParagraph) -> int:
        """
        Detect the heading level of a paragraph
        
        Args:
            paragraph: A docx paragraph
            
        Returns:
            int: The heading level (0 for normal paragraphs)
        """
        pass
    
    def _extract_text_formatting(self, paragraph: DocxParagraph) -> List[Dict[str, Any]]:
        """
        Extract text runs with their formatting
        
        Args:
            paragraph: A docx paragraph
            
        Returns:
            List[Dict]: List of runs with text and formatting info
        """
        pass 